feat: реализовать логику по друзьям

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -22,5 +22,5 @@ public class Film {
      * Продолжительность фильма в минутах
      */
     private Long duration;
-    private Set<Long> usersLiked;
+    private Set<Integer> usersLiked;
 }
